correct the number of free, no leaks - part 13

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -12,7 +12,7 @@
 size_t print_listint_safe(const listint_t *head)
 {
 	const listint_t *curr_node = head;
-	int number_of_nodes = 0;
+	size_t number_of_nodes = 0;
 
 	while (curr_node)
 	{
